+++ R/qm_simulation.R
@@ -506,20 +506,20 @@ get_1h_braino_basis_names <- function() {
 #' Note, this is a basic set and it may be appropriate to also include Asc, 
 #' Gly and PEth for high quality MRS data.
 #' 
-#' @param extra optional character vector of additional molecular names. Eg
+#' @param add optional character vector of additional molecular names. Eg
 #' c("asc", "gly", "peth").
 #' @param remove optional character vector of molecular names to remove from the
 #' set. Eg c("m_cr_ch2").
 #' @param inc_lip_mm include Lipid and MM basis signals.
 #' @return a character vector of molecule names.
 #' @export
-get_1h_brain_basis_names <- function(extra = NULL, remove = NULL,
+get_1h_brain_basis_names <- function(add = NULL, remove = NULL,
                                      inc_lip_mm = TRUE) {
   
   names <- c("m_cr_ch2", "ala", "asp", "cr", "gaba", "glc", "gln", "gsh", "glu",
              "gpc", "ins", "lac", "naa", "naag", "pch", "pcr", "sins", "tau")
   
-  if (!is.null(extra)) names <- sort(c(names, extra))
+  if (!is.null(add)) names <- sort(c(names, add))
   
   if (inc_lip_mm) {
     names <- c(names, "lip09", "lip13a", "lip13b", "lip20", "mm09", "mm12",